replaced wrong type
git-svn-id: http://svn.ratbox.org/svnroot/trunk@376 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/dbuf.c
@@ -149,7 +149,7 @@ int dbuf_put(struct DBuf* dyn, const char* buf, size_t length)
 {
   struct DBufBuffer** h;
   struct DBufBuffer*  db;
-  int                 chunk;
+  size_t              chunk;
 
   assert(0 != dyn);
   assert(0 != buf);
